v1.0.9
git-svn-id: https://plugins.svn.wordpress.org/sensei-buddypress/trunk@1267064 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/bp-sensei-group-settings.php
@@ -115,6 +115,11 @@ if ( class_exists( 'BP_Group_Extension' ) ) :
 			}
 			
 			if ( isset( $_POST[ 'bp_group_course' ] )  && ( $_POST[ 'bp_group_course' ] ) == '-1' ) {
+				
+				if ( !$old_course_id ) {
+					return;
+				}
+				
 				delete_post_meta($old_course_id, 'bp_course_group');
 				groups_delete_groupmeta( $group_id, 'bp_course_attached' );
 				bp_sensei_remove_members_group( $old_course_id, $group_id );
